Save and load alpha of conditional style color
Alpha (saturation) will be saved and loaded for the conditional style color of text and background.
Before there was not only the problem, that it was not saved - it was additional the issue, that background with no color had after reopening the project the alpha(saturation) of FF and the color black.
This is fixed now.